+++ src/main/java/com/team254/lib/trajectory/Spline.java
